requirements.txt
paho-mqtt==1.4.0
six==1.12.0
